chore(wrapper): rename `wrapper-ts` to `wrappers-ts`
