+++ docs/content/docs/acton-toml.mdx
@@ -171,7 +171,7 @@ generate-test = true
 test-output-dir = "tests"
 
 [wrappers.typescript]
-output-dir = "app/src/wrapper-ts"
+output-dir = "app/src/wrappers-ts"
 ```
 
 Override rules:
