+++ docs/content/docs/project-init.mdx
@@ -148,7 +148,7 @@ A simple example demonstrating state management in Tolk. It includes:
 
 The `counter` template also supports `--app`, which adds a React + Vite app,
 top-level `package.json` and `package-lock.json`, and a generated TypeScript
-wrapper in `wrapper-ts/Counter.ts`.
+wrapper in `wrappers-ts/Counter.ts`.
 
 ### `jetton`
 A comprehensive example of a [Jetton](https://docs.ton.org/develop/dapps/asset-processing/jettons) (Token) implementation. This is a great reference for building complex multi-contract applications. It includes:
