+++ docs/content/docs/scripting/blockchain-interaction.mdx
@@ -79,9 +79,9 @@ fun main() {
 }
 ```
 
-The `counter_wrapper.tolk` file provides a convenient wrapper around the counter contract:
+The `wrappers/Counter.tolk` file provides a convenient wrapper around the counter contract:
 
-```tolk collapse={2-3, 10-17, 30-42} title="counter_wrapper.tolk"
+```tolk collapse={2-3, 10-17, 30-42} title="wrappers/Counter.tolk"
 // todo update snippet to actual wrapper
 import "@acton/build"
 import "@acton/emulation/network"
