+++ src/commands/new/templates/counter-app/app/src/lib/counter.ts
@@ -2,7 +2,7 @@ import { Address, beginCell, storeStateInit, toNano } from '@ton/core';
 import { TonClient } from '@ton/ton';
 import type { SendTransactionParameters } from '@ton/appkit-react';
 
-import { Counter } from '../../../wrapper-ts/Counter';
+import { Counter } from '../../../wrappers-ts/Counter';
 import {
   IS_TESTNET,
   TONCENTER_API_KEY,
