+++ src/commands/new/templates/counter-app/tsconfig.json
@@ -17,5 +17,5 @@
     "jsx": "react-jsx",
     "types": ["vite/client"]
   },
-  "include": ["app/src", "wrapper-ts", "vite.config.ts"]
+  "include": ["app/src", "wrappers-ts", "vite.config.ts"]
 }
