+++ src/commands/new/templates/counter-app/wrappers-ts/Counter.ts
