+++ src/doc/man/generated_txt/acton-wrapper.txt
@@ -109,7 +109,7 @@ CONFIGURATION
            test-output-dir = "tests"
            
            [wrappers.typescript]
-           output-dir = "wrapper-ts"
+           output-dir = "wrappers-ts"
 
        CLI flags override config values for the current invocation.
 
@@ -137,11 +137,11 @@ EXAMPLES
 
               acton wrapper Counter --output-dir tests/generated
               acton wrapper Counter --test --test-output-dir tests/generated
-              acton wrapper Counter --ts --output-dir wrapper-ts
+              acton wrapper Counter --ts --output-dir wrappers-ts
 
        5. Generate a frontend-oriented TypeScript wrapper layout:
 
-              acton wrapper Counter --ts --output-dir app/src/wrapper-ts
+              acton wrapper Counter --ts --output-dir app/src/wrappers-ts
 
 SEE ALSO
        o  acton help build
