+++ src/etc/man/acton-new.1
@@ -333,7 +333,7 @@ When \fBacton new \-\-template counter \-\-app\fR is used, the project includes:
 .RE
 .sp
 .RS 4
-\h'-04'\(bu\h'+03'\fBwrapper\-ts/\fR for the generated TypeScript wrapper used by the app
+\h'-04'\(bu\h'+03'\fBwrappers\-ts/\fR for the generated TypeScript wrapper used by the app
 .RE
 .sp
 .RS 4
