+++ src/etc/man/acton-wrapper.1
@@ -138,7 +138,7 @@ generate\-test = true
 test\-output\-dir = "tests"
 
 [wrappers.typescript]
-output\-dir = "wrapper\-ts"
+output\-dir = "wrappers\-ts"
 .fi
 .RE
 .sp
@@ -192,7 +192,7 @@ acton wrapper Counter \-\-ts
 .nf
 acton wrapper Counter \-\-output\-dir tests/generated
 acton wrapper Counter \-\-test \-\-test\-output\-dir tests/generated
-acton wrapper Counter \-\-ts \-\-output\-dir wrapper\-ts
+acton wrapper Counter \-\-ts \-\-output\-dir wrappers\-ts
 .fi
 .RE
 .RE
@@ -202,7 +202,7 @@ acton wrapper Counter \-\-ts \-\-output\-dir wrapper\-ts
 .sp
 .RS 4
 .nf
-acton wrapper Counter \-\-ts \-\-output\-dir app/src/wrapper\-ts
+acton wrapper Counter \-\-ts \-\-output\-dir app/src/wrappers\-ts
 .fi
 .RE
 .RE
