+++ tests/integration/new_tests.rs
@@ -511,7 +511,7 @@ fn test_new_counter_project_with_app_flag() {
     ));
     assert!(!package_lock.contains(r#""name": "counter-project""#));
     assert!(project_dir.join("app/src/App.tsx").exists());
-    assert!(project_dir.join("wrapper-ts/Counter.ts").exists());
+    assert!(project_dir.join("wrappers-ts/Counter.ts").exists());
     assert!(project_dir.join("contracts/src/Counter.tolk").exists());
     assert!(
         project_dir
