+++ tests/snapshots/help_new/stdout.txt
@@ -241,7 +241,7 @@ COUNTER APP LAYOUT
 
        o  contracts/scripts for deployment and utility scripts
 
-       o  wrapper-ts/ for the generated TypeScript wrapper used by the app
+       o  wrappers-ts/ for the generated TypeScript wrapper used by the app
 
        o  app/ for the React + Vite frontend
